feat: begin auth implementation by adding mutations, resolvers, and server context prop
